@@ -1,68 +1,47 @@
 {
-  "name": "admin",
+  "name": "dashboard",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --configuration production",
-    "build:prod": "ng build --configuration production --base-href='/admin/'",
-    "lite-server": "lite-server --baseDir='dist'",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.4",
-    "@angular/common": "~12.0.4",
-    "@angular/compiler": "~12.0.4",
-    "@angular/core": "~12.0.4",
-    "@angular/forms": "~12.0.4",
-    "@angular/platform-browser": "~12.0.4",
-    "@angular/platform-browser-dynamic": "~12.0.4",
-    "@angular/router": "~12.0.4",
-    "@aws-amplify/ui-angular": "^1.0.9",
-    "@coreui/angular": "2.9.4",
-    "@coreui/coreui": "2.1.16",
-    "@coreui/coreui-plugin-chartjs-custom-tooltips": "1.3.1",
-    "@coreui/icons": "0.3.0",
-    "aws-amplify": "^4.0.2",
-    "bootstrap": "^4.5.0",
-    "chart.js": "^2.9.3",
-    "core-js": "^3.6.5",
-    "flag-icon-css": "^3.5.0",
-    "font-awesome": "^4.7.0",
-    "is-svg": "^4.2.2",
-    "lite-server": "^2.6.1",
-    "moment": "^2.27.0",
-    "mutationobserver-shim": "^0.3.7",
-    "ng2-charts": "^2.3.2",
-    "ngx-bootstrap": "^6.2.0",
-    "ngx-perfect-scrollbar": "^10.1.1",
-    "rxjs": "~6.6.0",
-    "simple-line-icons": "^2.5.2",
-    "ssri": "^8.0.1",
-    "tslib": "^2.0.0",
-    "web-animations-js": "^2.3.2",
+    "@angular/animations": "~14.0.0",
+    "@angular/cdk": "~14.0.4",
+    "@angular/common": "~14.0.0",
+    "@angular/compiler": "~14.0.0",
+    "@angular/core": "~14.0.0",
+    "@angular/flex-layout": "~14.0.0-beta.40",
+    "@angular/forms": "~14.0.0",
+    "@angular/material": "14.0.4",
+    "@angular/platform-browser": "~14.0.0",
+    "@angular/platform-browser-dynamic": "~14.0.0",
+    "@angular/router": "~14.0.0",
+    "@aws-amplify/ui-angular": "~2.4.14",
+    "aws-amplify": "~4.3.27",
+    "bootstrap": "~5.1.3",
+    "chart.js": "~3.8.0",
+    "chartjs-plugin-datalabels": "~2.0.0",
+    "ng2-charts": "~4.0.0",
+    "rxjs": "~7.5.0",
+    "tslib": "~2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.4",
-    "@angular/cli": "~12.0.4",
-    "@angular/compiler-cli": "~12.0.4",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.3",
+    "@angular-devkit/build-angular": "~14.1.0",
+    "@angular/cli": "~14.0.5",
+    "@angular/compiler-cli": "~14.0.0",
+    "@types/jasmine": "~4.0.0",
+    "jasmine-core": "~4.1.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.2.4"
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.7.2"
   }
 }
