@@ -2,7 +2,6 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: MIT-0
  */
-
 import { Injectable } from '@angular/core';
 import {
   HttpRequest,
@@ -12,27 +11,33 @@ import {
 } from '@angular/common/http';
 import { from, Observable } from 'rxjs';
 import { Auth } from 'aws-amplify';
-import { map, switchMap } from 'rxjs/operators';
+import { filter, map, switchMap } from 'rxjs/operators';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
   constructor() {}
-  idToken: string;
+  idToken = '';
 
   intercept(
     req: HttpRequest<any>,
     next: HttpHandler
   ): Observable<HttpEvent<any>> {
-    const session$ = from(Auth.currentSession());
+    if (req.url.includes('tenant/init')) {
+      return next.handle(req);
+    }
+    if (req.url.includes('auth-info') || req.url.includes('config.json')) {
+      return next.handle(req);
+    }
+
+    const s = Auth.currentSession().catch((err) => console.log(err));
+    const session$ = from(s);
 
     return session$.pipe(
-      map((sesh) => sesh.getIdToken()),
+      filter((sesh) => !!sesh),
+      map((sesh) => (!!sesh ? sesh.getIdToken().getJwtToken() : '')),
       switchMap((tok) => {
         req = req.clone({
-          headers: req.headers.set(
-            'Authorization',
-            'Bearer ' + tok.getJwtToken()
-          ),
+          headers: req.headers.set('Authorization', 'Bearer ' + tok),
         });
         return next.handle(req);
       })
