@@ -1,16 +1,7 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: MIT-0
- */
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-// Import Containers
-import { DefaultLayoutComponent } from './containers';
-
-import { P404Component } from './views/error/404.component';
-import { P500Component } from './views/error/500.component';
-import { UnauthorizedComponent } from './views/error/unauthorized.component';
+import { RouterModule, Routes } from '@angular/router';
+import { NavComponent } from './nav/nav.component';
+import { AuthComponent } from './views/auth/auth.component';
 
 export const routes: Routes = [
   {
@@ -19,30 +10,15 @@ export const routes: Routes = [
     pathMatch: 'full',
   },
   {
-    path: '404',
-    component: P404Component,
-    data: {
-      title: 'Page 404',
-    },
-  },
-  {
-    path: '500',
-    component: P500Component,
-    data: {
-      title: 'Page 500',
-    },
-  },
-  {
     path: '',
-    component: DefaultLayoutComponent,
+    component: NavComponent,
     data: {
       title: 'Home',
     },
     children: [
       {
-        path: 'auth',
-        loadChildren: () =>
-          import('./views/auth/auth.module').then((m) => m.AuthModule),
+        path: 'auth/info',
+        component: AuthComponent,
       },
       {
         path: 'dashboard',
@@ -54,20 +30,10 @@ export const routes: Routes = [
       {
         path: 'tenants',
         loadChildren: () =>
-          import('./tenants/tenant.module').then((m) => m.TenantModule),
-      },
-      {
-        path: 'users',
-        loadChildren: () =>
-          import('./users/users.module').then((m) => m.UsersModule),
+          import('./views/tenants/tenants.module').then((m) => m.TenantsModule),
       },
     ],
   },
-  {
-    path: 'unauthorized',
-    component: UnauthorizedComponent,
-  },
-  { path: '**', component: P404Component },
 ];
 
 @NgModule({
