@@ -3,9 +3,8 @@
  * SPDX-License-Identifier: MIT-0
  */
 import { Component, OnInit } from '@angular/core';
-import { getStyle, hexToRgba } from '@coreui/coreui/dist/js/coreui-utilities';
-import { CustomTooltips } from '@coreui/coreui-plugin-chartjs-custom-tooltips';
-import { TenantService } from '../../tenants/tenant.service';
+import { ChartConfiguration, ChartType } from 'chart.js';
+import { TenantService } from '../tenants/tenants.service';
 
 interface DataSet {
   label: string;
@@ -19,104 +18,59 @@ interface ChartData {
 
 @Component({
   templateUrl: 'dashboard.component.html',
+  styleUrls: ['./dashboard.component.scss'],
+  selector: 'app-dashboard',
 })
 export class DashboardComponent implements OnInit {
   constructor(private tenantSvc: TenantService) {}
 
   data: ChartData[] = [];
-  radioModel: string = 'Month';
-  public lineChart2Data: Array<any> = [
-    {
-      data: [1, 18, 9, 17, 34, 22, 11],
-      label: 'Series A',
-    },
-  ];
-  public lineChart2Labels: Array<any> = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-  ];
-  public lineChart2Options: any = {
-    tooltips: {
-      enabled: false,
-      custom: CustomTooltips,
-    },
-    maintainAspectRatio: false,
-    scales: {
-      xAxes: [
-        {
-          gridLines: {
-            color: 'transparent',
-            zeroLineColor: 'transparent',
-          },
-          ticks: {
-            fontSize: 2,
-            fontColor: 'transparent',
-          },
-        },
-      ],
-      yAxes: [
-        {
-          display: false,
-          ticks: {
-            display: false,
-            min: 1 - 5,
-            max: 50 + 5,
-          },
-        },
-      ],
-    },
-    elements: {
-      line: {
-        tension: 0.00001,
-        borderWidth: 1,
-      },
-      point: {
-        radius: 4,
-        hitRadius: 10,
-        hoverRadius: 4,
-      },
-    },
-    legend: {
-      display: false,
-    },
-  };
-  public lineChart2Colours: Array<any> = [
-    {
-      // grey
-      backgroundColor: getStyle('--info'),
-      borderColor: 'rgba(255,255,255,.55)',
-    },
-  ];
-  public lineChart2Legend = false;
-  public lineChart2Type = 'line';
-  // mainChart
 
-  public mainChartElements = 27;
-  public mainChartData1: Array<number> = [];
-  public mainChartData2: Array<number> = [];
-  public mainChartData3: Array<number> = [];
+  // lineChart
+  public lineChartElements = 27;
+  public lineChartData1: Array<number> = [];
+  public lineChartData2: Array<number> = [];
+  public lineChartData3: Array<number> = [];
 
-  public mainChartData: Array<any> = [
+  public lineChartData: Array<any> = [
     {
-      data: this.mainChartData1,
+      data: this.lineChartData1,
       label: 'Current',
+      backgroundColor: 'rgba(148,159,177,0.2)',
+      borderColor: 'rgba(148,159,177,1)',
+      pointBackgroundColor: 'rgba(148,159,177,1)',
+      pointBorderColor: '#fff',
+      pointHoverBackgroundColor: '#fff',
+      pointHoverBorderColor: 'rgba(148,159,177,0.8)',
+      fill: 'origin',
     },
     {
-      data: this.mainChartData2,
+      data: this.lineChartData2,
       label: 'Previous',
+      backgroundColor: 'rgba(77,83,96,0.2)',
+      borderColor: 'rgba(77,83,96,1)',
+      pointBackgroundColor: 'rgba(77,83,96,1)',
+      pointBorderColor: '#fff',
+      pointHoverBackgroundColor: '#fff',
+      pointHoverBorderColor: 'rgba(77,83,96,1)',
+      fill: 'origin',
     },
     {
-      data: this.mainChartData3,
+      data: this.lineChartData3,
       label: 'BEP',
+      backgroundColor: 'rgba(255,0,0,0.3)',
+      borderColor: 'red',
+      pointBackgroundColor: 'rgba(148,159,177,1)',
+      pointBorderColor: '#fff',
+      pointHoverBackgroundColor: '#fff',
+      pointHoverBorderColor: 'rgba(148,159,177,0.8)',
+      fill: 'origin',
     },
   ];
+
+  public lineChartLegend = false;
   /* tslint:disable:max-line-length */
-  public mainChartLabels: Array<any> = [
+  public lineChartLabels: Array<any> = [
     'Monday',
     'Tuesday',
     'Wednesday',
@@ -147,117 +101,43 @@ export class DashboardComponent implements OnInit {
     'Sunday',
   ];
   /* tslint:enable:max-line-length */
-  public mainChartOptions: any = {
-    tooltips: {
-      enabled: false,
-      custom: CustomTooltips,
-      intersect: true,
-      mode: 'index',
-      position: 'nearest',
-      callbacks: {
-        labelColor: function (tooltipItem, chart) {
-          return {
-            backgroundColor:
-              chart.data.datasets[tooltipItem.datasetIndex].borderColor,
-          };
-        },
-      },
-    },
-    responsive: true,
+
+  public lineChartType: ChartType = 'line';
+
+  public lineChartOptions: ChartConfiguration['options'] = {
     maintainAspectRatio: false,
-    scales: {
-      xAxes: [
-        {
-          gridLines: {
-            drawOnChartArea: false,
-          },
-          ticks: {
-            callback: function (value: any) {
-              return value.charAt(0);
-            },
-          },
-        },
-      ],
-      yAxes: [
-        {
-          ticks: {
-            beginAtZero: true,
-            maxTicksLimit: 5,
-            stepSize: Math.ceil(250 / 5),
-            max: 250,
-          },
-        },
-      ],
-    },
     elements: {
       line: {
-        borderWidth: 2,
-      },
-      point: {
-        radius: 0,
-        hitRadius: 10,
-        hoverRadius: 4,
-        hoverBorderWidth: 3,
+        tension: 0.5,
       },
     },
-    legend: {
-      display: false,
+    scales: {
+      // We use this empty structure as a placeholder for dynamic theming.
+      x: {},
+      'y-axis-0': {
+        position: 'left',
+      },
+      'y-axis-1': {
+        position: 'right',
+        grid: {
+          color: 'rgba(80,80,80,0.3)',
+        },
+        ticks: {
+          color: '#808080',
+        },
+      },
     },
   };
-  public mainChartColours: Array<any> = [
-    {
-      // brandInfo
-      backgroundColor: '#20a8d8',
-      borderColor: '#20a8d8',
-      pointHoverBackgroundColor: '#fff',
-    },
-    {
-      // brandSuccess
-      backgroundColor: 'transparent',
-      borderColor: getStyle('--success'),
-      pointHoverBackgroundColor: '#fff',
-    },
-    {
-      // brandDanger
-      backgroundColor: 'transparent',
-      borderColor: getStyle('--danger'),
-      pointHoverBackgroundColor: '#fff',
-      borderWidth: 1,
-      borderDash: [8, 5],
-    },
-  ];
-  public mainChartLegend = false;
-  public mainChartType = 'line';
 
   public random(min: number, max: number) {
     return Math.floor(Math.random() * (max - min + 1) + min);
   }
 
   ngOnInit(): void {
-    // generate random values for mainChart
-    this.tenantSvc.getTenants().subscribe((val) => {
-      val.forEach((tenant) => {
-        let simulatedOrders = [];
-        //simulate 7 months of orders
-        for (let j = 0; j < 7; j++) {
-          simulatedOrders.push(this.random(0, 50));
-        }
-        const dataSet: DataSet = {
-          label: 'Monthly Orders',
-          data: simulatedOrders,
-        };
-        this.data.push({
-          tenantId: tenant.tenantId,
-          dataSet: [dataSet],
-          totalOrders: simulatedOrders.reduce((a, b) => a + b, 0),
-        });
-      });
-    });
-
-    for (let i = 0; i <= this.mainChartElements; i++) {
-      this.mainChartData1.push(this.random(50, 200));
-      this.mainChartData2.push(this.random(80, 100));
-      this.mainChartData3.push(65);
+    for (let i = 0; i <= this.lineChartElements; i++) {
+      this.lineChartData1.push(this.random(50, 200));
+      this.lineChartData2.push(this.random(80, 100));
+      this.lineChartData3.push(65);
     }
   }
 }
