@@ -1,5 +1,5 @@
 export const environment = {
-  production: true,
+  production: false,
   clientId: '598qmpppgi26dqkoef4p2rrh3g',
   issuer: 'https://cognito-idp.us-west-2.amazonaws.com/us-west-2_eItw7wpUO',
   apiGatewayUrl: 'https://wchv3gaofb.execute-api.us-west-2.amazonaws.com/prod',
